+++ pczt/CHANGELOG.md
@@ -8,10 +8,11 @@ indicated by the `PLANNED` status in order to make it possible to correctly
 represent the transitive `semver` implications of changes within the enclosing
 workspace.
 
-## [Unreleased]
+## [0.7.0] - 2026-06-02
 
 ### Changed
-- Migrated to `zcash_protocol 0.9.0`, `zcash_transparent 0.8.0`, `zcash_primitives 0.28.0`, `zcash_proofs 0.28.0`.
+- Migrated to `zcash_protocol 0.9.0`, `zcash_transparent 0.8.0`, `zcash_primitives 0.28.0`,
+  `zcash_proofs 0.28.0`.
 
 ## [0.6.0] - 2026-04-27